Fix test that used a web page that changed
The page at https://underjord.io/live-server-push-without-js.html
changed and now the test fail. Use a simple embedded HTML file to test
the same condition (at least as it was described in the test).

It turns out, "fixing" this test actually exposes a bug in the
resolution of relative URIs. The whole resolution needs to be replaced
by `Uri.resolve()`, which will be done in subsequent commits.