Fix image relative URL resolution
The resolution of relative URLs in images is broken, as it is interpreting a
relative URL that doesn't start with `/` as an absolute path:

base URL: https://example.com/path/to/page.html
image URL: image.png

-> https://example.com/image.png

This is wrong and it should be resolved to https://example.com/path/to/image.png
instead. On top of that, other type of "relative" URLs, like "network-path
references" (URLs omiting only the scheme, like "//example.com/image.png") are
not handled at all.

Dart already provides URI resolution (Uri.resolve) as specified by RFC3986
(https://tools.ietf.org/html/rfc3986), so it is much more reliable to use that
instead of rolling our own.

This commit also supports having a null base URL in the Metadata, in which case
the image URL will simply not be resolved (the library client can perform their
own resolution if needed). A new test case is added to make sure this doesn't
break again.

Fixes #17.